% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getGroupName}
\alias{getGroupName}
\title{getGroupName}
\usage{
getGroupName(x)
}
\arguments{
\item{x}{Group name as it is downloaded, "_" between words}
}
\value{
The corresponding gene set group in the database.
}
\description{
This function is used to get the group name from downloaded gene groups using getGenesets function from EnrichmentBrowser.
}
\concept{Naming}
