% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getURL_ensembl}
\alias{getURL_ensembl}
\title{getURL_ensembl}
\usage{
getURL_ensembl(id)
}
\arguments{
\item{id}{ENSEMBL identifiers}
}
\value{
The corresponding URL's in the database.
}
\description{
It creates url from ensembl identifiers.
}
\seealso{
Other URL generation: 
\code{\link{getURL_WBC}()},
\code{\link{getURL_entrez}()}
}
\concept{URL generation}
