% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getURL_entrez}
\alias{getURL_entrez}
\title{getURL_entrez}
\usage{
getURL_entrez(id)
}
\arguments{
\item{id}{ENTREZ identifiers}
}
\value{
The corresponding URL's in the database.
}
\description{
It creates url from entrez identifiers.
}
\seealso{
Other URL generation: 
\code{\link{getURL_WBC}()},
\code{\link{getURL_ensembl}()}
}
\concept{URL generation}
