% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getURL_WBC}
\alias{getURL_WBC}
\title{getURL_WBC}
\usage{
getURL_WBC(id)
}
\arguments{
\item{id}{WBCgene identifiers}
}
\value{
The corresponding URL's in the database.
}
\description{
It creates url from WBCgene identifiers.
}
\seealso{
Other URL generation: 
\code{\link{getURL_ensembl}()},
\code{\link{getURL_entrez}()}
}
\concept{URL generation}
