% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gse69762}
\alias{gse69762}
\title{gse69762}
\format{
ExpressionSet
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE69762}
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-GEOD-69762/}
Paper/publication:
\url{https://www.ncbi.nlm.nih.gov/pubmed/26891258}
}
\description{
ExpressionSet \code{gse69762}.
Homo sapiens sapiens.
Expression profiling with microarrays.
Mucosal barrier dysfuction and its relation with Crohn's disease, CD,
pathogenesis. Further documentation can be found in the vignette
'T1_Microarray.Rmd'.
}
\examples{
data('gse69762', package = 'franciscomartinez')
}
\keyword{datasets}
