% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PRJNA601724}
\alias{PRJNA601724}
\title{PRJNA601724}
\format{
RangedSummarizedExperiment
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE143794}
 \url{https://www.ncbi.nlm.nih.gov/bioproject/601724}
}
\description{
SRA \code{SRP242209}
Caenorhabditis elegans
Expression profiling by high throughput sequencing.
RNAseq was carried out to compare the gene expression profiles of the worms
grown with individual bacteria to understand the bacterial effect. C. elegans
was grown with Chryseobacterium sp. CHNTR56, Comamonas sp. 12022 or
E. coli OP50.
}
\examples{
data(PRJNA601724,package='franciscomartinez')
}
\keyword{datasets}
